make qubes: force sign-all twice during the whole boot
It seems like specyfing sign-all in the middle of the list of targets,
and then at the end, didn't work, as make was skipping the seccond
occurance, thinking it already "built" it before. This resulted in
the 2nd stage packets (template, installer, etc) not being signed.

Additionally adding .PHONY marking for sign-all.