fedora: Clean yum metadata before installing build deps.
It is common to rebuild local package without actually changing the
version (so override rpm file). This makes yum metadata invalid - it
expects different file (as of content) under the same name, ending up
with marking package as invalid and trying to download again. The
actual error message is:

Error Downloading Packages:
  2000:kde-baseapps-4.9.5-1.fc18.x86_64: [Errno 256] No more mirrors to try.
  2000:kde-settings-plasma-4.9-19.1.fc18.noarch: [Errno 256] No more
mirrors to try.
  2000:kde-settings-ksplash-4.9-19.1.fc18.noarch: [Errno 256] No more
mirrors to try.
  2000:kde-baseapps-libs-4.9.5-1.fc18.x86_64: [Errno 256] No more
mirrors to try.

To fix this issue clear metadata before installing packages.