fedora chroot: do not exclude kernel package
Actually it doesn't hurt much here - this package can't be installed in
template, but in builder chroot it is just some waste of space.
Excluding 'kernel' package didn't solve any problem - 'kernel-debug' was
installed instead. And this was in conflict with
qubes-core-vm-kernel-placeholder.
So just accept the fact that kernel will be installed in chroot.