fedora chroot: exclude qubes-release package from local repo
This package will override fedora-release, which means that packages
would be build not in "Fedora" environment, but in "Qubes" environment,
which is undesired especially for template (VM) packages.
The only problem I've found so far is dist tag in release number (.qbs2
vs .fc18), but there can be more surprises.
Instead of excluding the package from command line (in every yum
call...) just add it to repository specification. Actually
'qubes-builder-pkgs' repo is enough, as other Qubes repositories are
disabled by default.