Introduce DISTS_VM_NO_FLAVOR to simplify non-templates rules
Most of the rules do not need template flavor, actually the first thing
is remove it if present. So to reduce code duplication, introduce
DISTS_VM_NO_FLAVOR which contains a list of bare dists, without
duplicates. Also use that in DISTS_ALL, for the same reason.

Bonus effect is that, if DISTS_VM contains multiple flavors of the same
distribution, the component packages will be built only once.