Rework update-repo-* targets
1. Use makefile targets instead of shell loops - this make the actual
   shell code much more readable (at some cost of readability of the loops
   itself, but that's explained in comments)
2. Unify update-repo-current with the rest of update-repo-* targets,
   instead of hack we had previously.
3. Improve documentation on builder plugins API in that respect.

General idea is to use `update-repo` target of builder plugin to include
packages in every but 'current' repository type and store package list
in `SNAPSHOT_FILE` - this is unchanged. But then use
`update-repo-from-snapshot` target from the same plugin to actually
parse `SNAPSHOT_FILE` and include the packages in target repository
('current'). Previously the last step was implemented in separate script
in totally different repository, with hardcoded paths.

This all additionally make MIN_AGE variable easier to find and possibly
change. And also makes it possible to hook additional plugins to
`update-repo-from-snapshot` target (having `builder-github` in mind).