get-sources: Use canonical make facilities to simplify
As part of this change, remove the undocumented dependency on $SCRIPT_DIR
in get-sources as well as the special casing for fetching builder plugins
first - it serves no purpose.

The most obvious benefit is that parallel building of this target now
works, i.e. 'make get-sources -jN' will fetch the sources in parallel,
and is much faster here.