Make the top level signing targets empty rules
This is a trial change to get feedback. The use of @true in otherwise
empty rules prevents make from determining that nothing needs to be done.
It is clearer and more concise to express this as an empty rule. It also
allows make to report that a target is up to date vs silent execution of
what may be a no-op (but may also be some quickish non-verbose stuff the
user doesn't know about).