update-repo-*: don't try to upload packages if component doesn't specify any
In theory every builder plugin implement similar check itself, but it
makes more sense to not call update-repo-* in such a case at all.

This is especially important for update-repo-current, which have
additional date check (if packages were in testing long enough). This
check fails the whole update-repo call if no packages are defined for a
given target (like vm/dom0).