qubesbuilder.BuildLog: add support for builder-controlled subdir
Ease logs organization - allow a sender (builder) to specify a single
subdir to place the log into. This is supposed to be a hint what this
build is about (package name, distribution version etc). Since this is
send as a service call argument, the build process itself can't control
it, as it it send before the build starts - so a compromise during the
build cannot fake it.