Take advantage of git tag --format
`git-tag(1)` supports the option `--format`, which allows it to list
information other than just the tag names.  In particular, it can list
the tag objectname (hash) and object (hash the tag points to).  By
parsing the output of `git tag '--format=%(objectname):%(object):'`, the
need to convert tag names to hashes is eliminated.  Furthermore, since
the hash the tag points to is available, comparing it to the expected
hash is trivial.

I also used this opportunity to perform additional refactoring.  The
code that verifies git object signatures has been moved into a single
shell function, and the exit path has been simplified.