scripts: Avoid using 'git branch --show-current'
The 'show-current' option was introduced in git 2.22. As the standard Fedora
template in Qubes 4.0 still ships version 2.21.3 by default 'make get-sources' 
fails on these systems.

This commit works around this issue by replacing the command for retrieving the
current branch name with one which is also supported by older versions of git.