Fix parsing template options with empty flavor
This is relevant for example with 'selinux' template option, but with
default flavor: "fc37++selinux" should be parsed as "fc37" dist, empty
flavor and "selinux" option. Previously it was parsed as "selinux"
flavor, which doesn't exist. Use 'read -a' bash builtin instead of
string replace to construct array, without collapsing repeated
delimiters.

Related to QubesOS/qubes-issues#7988